# swagger/swaggerapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Contacts API",
    "version": "1.0.0",
    "description": "CSE 341 – W02 Contacts CRUD"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    },
    {
      "url": "https://cse341-contacts-5rf5.onrender.com"
    }
  ],
  "paths": {
    "/contacts": {
      "get": {
        "summary": "GET all contacts",
        "tags": [
          "Contacts"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Contact"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "POST create a contact (all fields required)",
        "tags": [
          "Contacts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing fields"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "summary": "GET by id",
        "tags": [
          "Contacts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "PUT update (replace all fields)",
        "tags": [
          "Contacts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactCreate"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Updated"
          },
          "400": {
            "description": "Invalid id or missing fields"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "DELETE contact",
        "tags": [
          "Contacts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "400": {
            "description": "Invalid id"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Contact": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "665ee0f8fb25e2a629fa2d11"
          },
          "firstName": {
            "type": "string",
            "example": "Ada"
          },
          "lastName": {
            "type": "string",
            "example": "Lovelace"
          },
          "email": {
            "type": "string",
            "example": "ada@example.com"
          },
          "favoriteColor": {
            "type": "string",
            "example": "Teal"
          },
          "birthday": {
            "type": "string",
            "example": "1815-12-10"
          }
        }
      },
      "ContactCreate": {
        "type": "object",
        "required": [
          "firstName",
          "lastName",
          "email",
          "favoriteColor",
          "birthday"
        ],
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "favoriteColor": {
            "type": "string"
          },
          "birthday": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Contacts"
    }
  ]
}